♻️ : #467 - ExplorePlans 비즈니스 로직 분리
