+++ src/app/(header)/explore/_components/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useTab } from './useTab';
+export { useExplorePlans } from './useExplorePlans';
